requirements.txt
torch==2.10.0
torchvision==0.25.0
streamlit==1.45.0
numpy==2.1.0
matplotlib==3.9.0
Pillow==10.4.0
scikit-learn==1.5.0